// frontend/src/books/types/Book.ts
export default interface Book {
    _pt: string;
    _abstract: string;
    _annote: number;
    _author: string;
    _doi: string;
    _issn: number;
    _journal: string;
    _keywords: string;
    _month: string;
    _pages: string;
    _publisher: string;
    _title: string;
    _url: string;
    _volume: number;
    _year: number;
    _id: string;
}